+++ src/App.js
@@ -1,5 +1,5 @@
-const OutputView = require('./OutputView');
-const InputView = require('./InputView');
+const OutputView = require('./View/OutputView');
+const InputView = require('./View/InputView');
 
 class App {
   play() {
