+++ src/View/InputView.js
@@ -1,16 +1,17 @@
 const MissionUtils = require('@woowacourse/mission-utils');
-const BridgeMaker = require('./BridgeMaker');
-const BridgeRandomNumberGenerator = require('./BridgeRandomNumberGenerator');
-const BridgeGame = require('./BridgeGame');
-const Check = require('./Check');
-const BridgePrint = require('./BridgePrint');
+const BridgeMaker = require('../BridgeMaker');
+const BridgeRandomNumberGenerator = require('../BridgeRandomNumberGenerator');
+const BridgeGame = require('../BridgeGame');
+const Check = require('../Utils/Check');
+const Vaild = require('../Utils/Vaild');
+const BridgePrint = require('../BridgePrint');
 
-const { INPUT_MESSAGES } = require('./common/messages');
+const { INPUT_MESSAGES } = require('../common/messages');
 const InputView = {
   readBridgeSize() {
     MissionUtils.Console.readLine(`${INPUT_MESSAGES.INPUT_BRIDGESIZE}`, (size) => {
       MissionUtils.Console.print('');
-      Check.isVaildSize(size);
+      Vaild.isVaildSize(size);
       const brige = BridgeMaker.makeBridge(size, BridgeRandomNumberGenerator.generate);
       const gamePlay = new BridgeGame(brige);
       this.readMoving(gamePlay);
@@ -19,7 +20,7 @@ const InputView = {
 
   readMoving(gamePlay) {
     const playerInput = (input) => {
-      Check.isVaildInput(input);
+      Vaild.isVaildInput(input);
       gamePlay.move(input);
       if (this.checkGameOver(gamePlay) === 0) return;
       return this.readMoving(gamePlay);
